requirements-test.txt
Flask==1.1.1
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
